% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_wrappers.R
\name{is_cpp_enabled}
\alias{is_cpp_enabled}
\title{Check if C++ Acceleration is Enabled}
\usage{
is_cpp_enabled()
}
\value{
Logical indicating if C++ acceleration is currently enabled
}
\description{
Check if C++ Acceleration is Enabled
}
