% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphericalSIM-package.R
\docType{package}
\name{sphericalSIM-package}
\alias{sphericalSIM}
\alias{sphericalSIM-package}
\title{sphericalSIM: Spherical Single-Index Models}
\description{
Spherical single-index models with adaptive group lasso variable selection.
}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
\keyword{internal}
